Parser: Allow more flexibility and combine pkg resolving
