+++ lib/initialize.js
@@ -2,7 +2,7 @@ const questions = require('./utils/observable-questions');
 const validateAddons = require('./utils/validate-addons');
 const inq = require('./inquirer-prompt');
 
-//TODO: add with flag to check if there could be a dependency beneath later
+//TODO: allow to add question to the call
 module.exports = function initializeInquirer(pkg) {
 	if(pkg.length == 0) {
 		let config = {
@@ -10,7 +10,8 @@ module.exports = function initializeInquirer(pkg) {
 			output: ''
 		};
 		inq(questions, config);
-	} else {
+	}
+	else {
 		validateAddons(pkg);
 	}
 };
