+++ lib/inquirer-prompt.js
@@ -1,6 +1,7 @@
 const inquirer = require('inquirer');
 const parse = require('./parser/index');
 const questions = require('./utils/observable-questions');
+const chalk = require('chalk');
 
 module.exports = function(questions, config) {
 
@@ -29,12 +30,19 @@ function attachAnswers(config, ans) {
 }
 
 function checkEmptyAnswers(config) {
+	let errors = [];
 	for(let key in config) {
 		if(!config[key]) {
-			console.error('\nFound no answer given to property', key);
-			console.error('Check if you have the same name for the inquirer property!');
-			process.exit(0);
+			errors[key] = chalk.gray.bold('\nFound no answer given to property ') +
+			chalk.red.bold(key) + '\n';
 		}
 	}
+	// get errors and use filter to discard the process after we've logged each error
+	Object.keys(errors).map( (err) => {
+		console.error(errors[err]);
+	}).filter( () => {
+		process.exit(0);
+	});
+
 	return;
 }
