+++ lib/parser/index.js
@@ -1,15 +1,13 @@
-const resolvePackages = require('./resolve-packages');
 const resolveTransform = require('./resolve-transform');
-//eslint-disable-next-line
 const validateOptions = require('./validateOptions');
 
+//TODO: Allow options and pkg to be passed down.
 module.exports = function parser(pkg,opts) {
 	// null, config -> without package
 	// addon, null -> with package
+	console.log(pkg, opts, 'Hello');
 	if(opts) {
 		validateOptions(opts);
-	} else if(!opts) {
-		resolvePackages(pkg);
 	} else {
 		resolveTransform();
 	}
