+++ lib/parser/resolve-transform.js
@@ -1,4 +1,5 @@
-// Insert our own rules here and send it to the parser on the init option
-module.exports = function resolveTransform(option) { //eslint-disable-line
+//TODO: Get the inq prompt and return the parser once again until there's no packages left to validate.
 
+module.exports = function resolveTransform(option) { //eslint-disable-line
+	console.log('resolved');
 };
