+++ lib/utils/validate-addons.js
@@ -1,5 +1,5 @@
 const exists = require('./npm-exists');
-const parse = require('../parser/index');
+const resolvePackages = require('./resolve-packages');
 const chalk = require('chalk');
 
 module.exports = function validateAddons(addon) {
@@ -15,7 +15,9 @@ module.exports = function validateAddons(addon) {
 			if(!moduleExists) {
 				throw new TypeError('Package isn\'t registered on npm.');
 			}
-			if (moduleExists) parse(addon, null);
+			if (moduleExists) {
+				return resolvePackages(pkg);
+			}
 		}).catch(err => {
 			console.error(err.stack || err);
 			process.exit(0);
